refactor: move collect schema coordinates file to dedicated file and add some tests
