+++ packages/libraries/core/src/client/usage.ts
@@ -1,31 +1,16 @@
 import {
-  ArgumentNode,
-  ASTNode,
   DocumentNode,
-  getNamedType,
-  GraphQLInputType,
-  GraphQLInterfaceType,
-  GraphQLNamedType,
-  GraphQLObjectType,
   GraphQLSchema,
-  GraphQLType,
-  GraphQLUnionType,
-  isEnumType,
-  isInputObjectType,
-  isScalarType,
   Kind,
-  NameNode,
-  ObjectFieldNode,
   OperationDefinitionNode,
   TypeInfo,
-  visit,
-  visitWithTypeInfo,
   type ExecutionArgs,
 } from 'graphql';
 import LRU from 'tiny-lru';
 import { normalizeOperation } from '../normalize/operation.js';
 import { version } from '../version.js';
 import { createAgent } from './agent.js';
+import { collectSchemaCoordinates } from './collect-schema-coordinates.js';
 import { dynamicSampling, randomSampling } from './sampling.js';
 import type {
   AbortAction,
@@ -322,258 +307,13 @@ export function createCollector({
       [key: string]: unknown;
     } | null,
   ): CacheResult {
-    const entries = new Set<string>();
-    const collected_entire_named_types = new Set<string>();
-    const shouldAnalyzeVariableValues = processVariables === true && variables !== null;
-
-    function markAsUsed(id: string) {
-      if (!entries.has(id)) {
-        entries.add(id);
-      }
-    }
-
-    function makeId(...names: string[]): string {
-      return names.join('.');
-    }
-
-    const collectedInputTypes: Record<
-      string,
-      {
-        all: boolean;
-        fields: Set<string>;
-      }
-    > = {};
-
-    function collectInputType(inputType: string, fieldName?: string) {
-      if (!collectedInputTypes[inputType]) {
-        collectedInputTypes[inputType] = {
-          all: false,
-          fields: new Set<string>(),
-        };
-      }
-
-      if (fieldName) {
-        collectedInputTypes[inputType].fields.add(fieldName);
-      } else {
-        collectedInputTypes[inputType].all = true;
-      }
-    }
-
-    function collectNode(node: ObjectFieldNode | ArgumentNode) {
-      const inputType = typeInfo.getInputType()!;
-      const inputTypeName = resolveTypeName(inputType);
-
-      if (node.value.kind === Kind.ENUM) {
-        // Collect only a specific enum value
-        collectInputType(inputTypeName, node.value.value);
-      } else if (node.value.kind !== Kind.OBJECT && node.value.kind !== Kind.LIST) {
-        // When processing of variables is enabled,
-        // we want to skip collecting full input types of variables
-        // and only collect specific fields.
-        // That's why the following condition is added.
-        // Otherwise we would mark entire input types as used, and not granular fields.
-        if (node.value.kind === Kind.VARIABLE && shouldAnalyzeVariableValues) {
-          return;
-        }
-
-        collectInputType(inputTypeName);
-      }
-    }
-
-    function markEntireTypeAsUsed(type: GraphQLInputType): void {
-      const namedType = getNamedType(type);
-
-      if (collected_entire_named_types.has(namedType.name)) {
-        // No need to mark this type as used again
-        return;
-      }
-      // Add this type to the set of types that have been marked as used
-      // to avoid infinite loops
-      collected_entire_named_types.add(namedType.name);
-
-      if (isScalarType(namedType)) {
-        markAsUsed(makeId(namedType.name));
-        return;
-      }
-
-      if (isEnumType(namedType)) {
-        namedType.getValues().forEach(value => {
-          markAsUsed(makeId(namedType.name, value.name));
-        });
-        return;
-      }
-
-      const fieldsMap = namedType.getFields();
-
-      for (const fieldName in fieldsMap) {
-        const field = fieldsMap[fieldName];
-
-        markAsUsed(makeId(namedType.name, field.name));
-        markEntireTypeAsUsed(field.type);
-      }
-    }
-
-    function collectVariable(namedType: GraphQLNamedInputType, variableValue: unknown) {
-      const variableValueArray = Array.isArray(variableValue) ? variableValue : [variableValue];
-      if (isInputObjectType(namedType)) {
-        variableValueArray.forEach(variable => {
-          if (variable) {
-            // Collect only the used fields
-            for (const fieldName in variable) {
-              const field = namedType.getFields()[fieldName];
-              if (field) {
-                collectInputType(namedType.name, fieldName);
-                collectVariable(getNamedType(field.type), variable[fieldName]);
-              }
-            }
-          } else {
-            // Collect type without fields
-            markAsUsed(makeId(namedType.name));
-          }
-        });
-      } else {
-        collectInputType(namedType.name);
-      }
-    }
-
-    visit(
-      doc,
-      visitWithTypeInfo(typeInfo, {
-        Field(node, _key, _parent, path, ancestors) {
-          const parent = typeInfo.getParentType();
-          const field = typeInfo.getFieldDef();
-
-          if (!parent) {
-            throw new Error(
-              `Could not find a parent type of a field at ${printPath(path, ancestors, node.name)}`,
-            );
-          }
-
-          if (!field) {
-            throw new Error(
-              `Could not find a field definition of a field at ${printPath(
-                path,
-                ancestors,
-                node.name,
-              )}`,
-            );
-          }
-
-          markAsUsed(makeId(parent.name, field.name));
-        },
-        VariableDefinition(node) {
-          const inputType = typeInfo.getInputType();
-
-          if (!inputType) {
-            throw new Error(`Could not find an input type of a variable $${node.variable.name}`);
-          }
-
-          if (shouldAnalyzeVariableValues) {
-            // Granular collection of variable values is enabled
-            const variableName = node.variable.name.value;
-            const variableValue = variables[variableName];
-            const namedType = getNamedType(inputType);
-
-            collectVariable(namedType, variableValue);
-          } else {
-            // Collect the entire type without processing the variables
-            collectInputType(resolveTypeName(inputType));
-          }
-        },
-        Directive(node) {
-          return {
-            ...node,
-            arguments: [],
-          };
-        },
-        Argument(node, _key, _parent, path, ancestors) {
-          const parent = typeInfo.getParentType();
-          const field = typeInfo.getFieldDef();
-          const arg = typeInfo.getArgument();
-
-          if (!parent) {
-            throw new Error(
-              `Could not find a parent type of an argument at ${printPath(
-                path,
-                ancestors,
-                node.name,
-              )}`,
-            );
-          }
-
-          if (!field) {
-            throw new Error(
-              `Could not find a field definition of an argument at ${printPath(
-                path,
-                ancestors,
-                node.name,
-              )}`,
-            );
-          }
-
-          if (!arg) {
-            throw new Error(
-              `Could not find an argument definition of an argument at ${printPath(
-                path,
-                ancestors,
-                node.name,
-              )}`,
-            );
-          }
-
-          markAsUsed(makeId(parent.name, field.name, arg.name));
-          collectNode(node);
-        },
-        ListValue(node, _key, _parent, path, ancestors) {
-          const inputType = typeInfo.getInputType();
-
-          if (!inputType) {
-            throw new Error(
-              `Could not find an input type of a list value at ${printPath(path, ancestors)}`,
-            );
-          }
-
-          const inputTypeName = resolveTypeName(inputType);
-
-          node.values.forEach(value => {
-            if (value.kind !== Kind.OBJECT) {
-              // if a value is not an object we need to collect all fields
-              collectInputType(inputTypeName);
-            }
-          });
-        },
-        ObjectField(node, _key, _parent, path, ancestors) {
-          const parentInputType = typeInfo.getParentInputType();
-
-          if (!parentInputType) {
-            throw new Error(
-              `Could not find an input type of an object field at ${printPath(
-                path,
-                ancestors,
-                node.name,
-              )}`,
-            );
-          }
-
-          const parentInputTypeName = resolveTypeName(parentInputType);
-
-          collectNode(node);
-          collectInputType(parentInputTypeName, node.name.value);
-        },
-      }),
-    );
-
-    for (const inputTypeName in collectedInputTypes) {
-      const { fields, all } = collectedInputTypes[inputTypeName];
-
-      if (all) {
-        markEntireTypeAsUsed(schema.getType(inputTypeName) as any);
-      } else {
-        fields.forEach(field => {
-          markAsUsed(makeId(inputTypeName, field));
-        });
-      }
-    }
+    const entries = collectSchemaCoordinates({
+      documentNode: doc,
+      processVariables,
+      schema,
+      typeInfo,
+      variables,
+    });
 
     return {
       document: normalizeOperation({
@@ -593,43 +333,6 @@ export function createCollector({
     LRU<CacheResult>(max, ttl),
   );
 }
-
-function resolveTypeName(inputType: GraphQLType): string {
-  return getNamedType(inputType).name;
-}
-
-function printPath(
-  path: readonly (string | number)[],
-  ancestors: readonly (ASTNode | readonly ASTNode[])[],
-  leafNameNode?: NameNode,
-): string {
-  const result: string[] = [];
-  for (let i = 0; i < path.length; i++) {
-    const key = path[i];
-    const ancestor = ancestors[i];
-
-    if (!ancestor) {
-      break;
-    }
-
-    if (key === 'selectionSet') {
-      if ('name' in ancestor && ancestor.name?.value) {
-        result.push(ancestor.name.value);
-      }
-    }
-  }
-
-  if (leafNameNode) {
-    result.push(leafNameNode.value);
-  }
-
-  return result.join('.');
-}
-
-type GraphQLNamedInputType = Exclude<
-  GraphQLNamedType,
-  GraphQLObjectType | GraphQLInterfaceType | GraphQLUnionType
->;
 
 export interface Report {
   size: number;
